HoptArgData

Description

The configuration for Hopt and lenses to access each field. This small module is kept separate from the rest, because it uses TemplateHaskell to generate the Lens instances.

Documentation

data Cfg

Configuration data for Hopt

Constructors

Cfg

Fields

_optPasses :: [String]

List of optimization passes

_inFile :: FilePath

Input file

_outFile :: FilePath

Output file

_isText :: Bool

Output as text, not bitcode

outFile :: Lens Cfg Cfg FilePath FilePath

optPasses :: Lens Cfg Cfg [String] [String]

isText :: Lens Cfg Cfg Bool Bool

inFile :: Lens Cfg Cfg FilePath FilePath