LlvmParser

Description

A parser for a very small subset of LLVM. What is unique about these parser combinators is that they parse whitespace preceding each token instead of after them. This means the parser will return as soon as possible, and not hang waiting for whitespace.

Documentation

basicBlock :: Parser Block

Parses a basic block

basicBlock' :: Parser Block

Parses a statement followed by a terminator or a terminator

stat :: Parser Statement

Parses a statement

branchStat :: Parser Statement

Parses an unconditional branch statement

branchCondStat :: Parser Statement

Parses a conditional branch statement

boolLit :: Parser Literal

Parses a Boolean literal

flushStat :: Parser Statement

Parses a special flush statement, used to flush the parser pipeline

terminator :: Parser ()

Parses a statement terminator

isHorizSpace :: (Enum a, Eq a) => a -> Bool

Returns true if the input character is a tab or space

eord :: Enum a => Char -> a

Same as ord, but returns any Enum instead of an Int

isTerminator :: (Enum a, Eq a) => a -> Bool

Returns true if the input character is a terminator

comment :: Parser ByteString

Parses a comment

assignStat :: Parser Statement

Parses an assignment statement

returnStat :: Parser Statement

Parses a return statement

expression :: Parser Expr

Parses an expression

addExpr :: Parser Expr

Parses an add expression

phiExpr :: Parser Expr

Parses a phi expression

phiField :: Parser (Expr, String)

Parses a phi field

integer :: Parser Integer

Parses an integer

globalVar :: Parser String

Parses a global variable

identifier :: Parser String

Parses an identifier

labelStat :: Parser Statement

Parses a label

isLetterOrUnder :: (Enum a, Ord a) => a -> Bool

Parses a letter or underscore

isAlphaNumOrUnder :: (Enum a, Ord a) => a -> Bool

Parses an alphanumeric character or an underscore

isNum :: (Enum a, Ord a) => a -> Bool

Returns true if the input character is a number

keyword :: ByteString -> Parser ()

Requires some whitespace after keyword

token :: ByteString -> Parser ()

Optional whitespace on either side of a given string

whitespace :: Parser ()

Parse zero or more whitespace characters

whitespace1 :: Parser ()

Parse one or more whitespace characters

isWhite :: (Enum a, Eq a) => a -> Bool

Returns true if the input character is a whitespace character

toplevelEntities :: Parser Module

Parses a list of top-level entities

toplevelEntity :: Parser ToplevelEntity

Parses a top-level entity

target :: Parser ToplevelEntity

Parses a target definition

targetName :: Parser String

Parses a target name

stringLiteral :: Parser String

Parses a string literal

function :: Parser ToplevelEntity

Parses a function definition

parens :: Parser a -> Parser a

Parses p within parens

braces :: Parser a -> Parser a

Parses p within braces

brackets :: Parser a -> Parser a

Parses p within brackets

argument :: Parser String

Parses a function argument

comma :: Parser ()

Parses a comma

llvmType :: Parser String

Parses an LLVM type

intType :: Parser String

Parses an integer type

functionAttribute :: Parser String

Parses a function attribute