LlvmPrinter

Description

A pretty-printer for a small subset of LLVM

Documentation

class ToLlvm a where

A typeclass for transforming any type to any monoidal string type

Methods

toLlvm :: (Eq s, Monoid s, IsString s) => a -> s

bbLine :: (Monoid s, IsString s) => s -> s

Print a basic block line

identifier :: (Monoid s, IsString s) => String -> s

Print an identifier

mintercalate :: (Monoid s, IsString s) => s -> [s] -> s

Same as intercalate from Data.List, but works for any monoid

phiField :: (Eq s, Monoid s, IsString s) => (Expr, String) -> s

Prints a phi field

(<+>) :: (Eq s, Monoid s, IsString s) => s -> s -> s

Concat with a space between, unless one is empty