Main

Description

An implementation of a very small subset of LLVM's opt

Documentation

main :: IO ()

The entry point for hopt

compile :: Cfg -> IO ()

Given a configuration, compile the input file and write the output file

input :: FilePath -> Onum ByteString IO a

The input pipe that enumerates the input file. If p is a single hyphen, stdin is used.

output :: FilePath -> IO (Iter ByteString IO ())

The output iterator that writes the output file. If p is a single hyphen, stdout is used.