Transforms.DeadInstructionElimination

Description

Dead Instruction Elimination

This optimization pass removes assignment statements that are unreferenced by all later statements. Unlike the LLVM version, this version also moves each instruction as closely as possible to the first instruction that references it.

How it works:

When a definition (i.e. assignment) is parsed, we pull it out of the basic block until another statement references it. Then, for each statement, we lookup all definitions it uses and check for them in our list of unreferenced definitions. If we find one, we push the definition back into the basic block, just before the statement that references it.

For example:

Line Input PassState Output Comment ---- -------- --------- ------ ------- 1 %1 = 5 [%1 = 5] %1 may be unreferrenced 2 %2 = 7 [%1 = 5, %2 = 7] %1 & %2 may be unreferrenced 3 %3 = %1 [%2 = 7, %3 = %1] [%1 = 5] %1 is used 4 ret %3 [%2 = 7] [%3 = %1, ret %3] %3 is used, %2 is dropped

Limitations:

If there is a chain of unreferenced definitions, this pass will only remove the first link in the chain. 2. Code size: A redundant definition can be pulled into branch if both sides reference the definition for the first time.

Hindsight:

This pass is probably more useful in improving spacial locality and reducing register pressure than removing dead instructions.

Documentation

type PassState = [(String, Expr)]

A list of unreferenced instructions. The String is a variable name, and the Expr is the value assigned to it.

name :: String

The name of this optimization pass

emptyState :: PassState

The initial state

chunk :: Monad m => Module -> StateT PassState m Module

Optimize a list of top-level entities

toplevelEntity :: Monad m => ToplevelEntity -> StateT PassState m ToplevelEntity

Optimize a top-level entity

stat :: Monad m => Statement -> StateT PassState m Block

Optimize a statement

vars :: Expr -> [String]

Given an expression, return the variable names it references

vars' :: Expr -> [String]

Given an expresion, if it is variable, return its name. Otherwise, return an empty list

pushStatement :: Monad m => String -> Expr -> StateT PassState m ()

Push a statement that may be dead

popStatement :: Monad m => String -> StateT PassState m [Statement]

Lookup a statement that is now known to be undead. Remove it from the list of dead instructions.