Transforms.DeadInstructionElimination

Description

Dead Instruction Elimination

This optimization pass removes assignment statements that are unreferenced by all later statements. Unlike the LLVM version, this version also moves each instruction as closely as possible to the first instruction that references it.

How it works:

When a definition (i.e. assignment) is parsed, we pull it out of the basic block until another statement references it. Then, for each statement, we lookup all definitions it uses and check for them in our list of unreferenced definitions. If we find one, we push the definition back into the basic block, just before the statement that references it.

For example:

Line Input PassState Output Comment ---- -------- --------- ------ ------- 1 %1 = 5 [%1 = 5] %1 may be unreferrenced 2 %2 = 7 [%1 = 5, %2 = 7] %1 & %2 may be unreferrenced 3 %3 = %1 [%2 = 7, %3 = %1] [%1 = 5] %1 is used 4 ret %3 [%2 = 7] [%3 = %1, ret %3] %3 is used, %2 is dropped

Limitations:

If there is a chain of unreferenced definitions, this pass will only remove the first link in the chain. 2. Code size: A redundant definition can be pulled into branch if both sides reference the definition for the first time.

Hindsight:

This pass is probably more useful in improving spacial locality and reducing register pressure than removing dead instructions.

Documentation

type PassState = [(String, Expr)]

name :: String

emptyState :: PassState

chunk :: MonadState PassState m => Module -> m Module

toplevelEntity :: MonadState PassState m => ToplevelEntity -> m ToplevelEntity

stat :: MonadState PassState m => Statement -> m Block

vars :: Expr -> [String]

vars' :: Expr -> [String]

pushStatement :: MonadState PassState m => String -> Expr -> m ()

popStatement :: MonadState PassState m => String -> m [Statement]